samples/helm/device-registry-app/requirements.txt
Flask==2.0.3
Flask-RESTful==0.3.9
requests==2.27.1
pytest==7.0.1
mysql-connector-python==8.0.28
markdown==3.3.7
